+++ CHANGELOG.md
@@ -21,6 +21,8 @@ CHANGED browser.referer -> browser.referrer
 
 CHANGED removed onalert/onconfirm/onprompt, use browser.on('alert', fn) etc
 
+CHANGED credentials replaced with browser.on('authenticate', fn)
+
 CHANGED browser.wait() runs event loop so you can catch done/error events
 
 CHANGED event loop `done` event renamed to `idle`
@@ -29,7 +31,7 @@ CHANGED upgraded to JSDOM 3.1, added Node 0.12 and io.js support
 
 FIXED XHR errors no longer stop the event loop
 
- 694 passing (25s)
+ 690 passing (25s)
   12 pending